+++ package.json
@@ -27,12 +27,12 @@
     "mammoth": "^1.9.0",
     "nepali-date-converter": "^3.3.2",
     "net": "^1.0.2",
-    "next": "^14.2.1",
+    "next": "^15.3.5",
     "next-mdx-remote": "^4.4.1",
     "nodemailer": "^6.10.0",
-    "react": "^18",
+    "react": "^19.1.0",
     "react-confetti": "^6.4.0",
-    "react-dom": "^18",
+    "react-dom": "^19.1.0",
     "react-firebase-hooks": "^5.1.1",
     "react-icons": "^5.5.0",
     "react-papaparse": "^4.4.0",
