+++ tsconfig.json
@@ -1,6 +1,10 @@
 {
   "compilerOptions": {
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -18,9 +22,27 @@
       }
     ],
     "paths": {
-      "@/*": ["./src/*"]
-    }
+      "@/*": [
+        "./src/*"
+      ]
+    },
+    "target": "ES2017"
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/app/Resource/[id]/page.jsx", "src/app/Resources/page.jsx", "src/app/layout.tsx", "src/app/page.jsx", "src/app/Members/page.jsx", "src/app/Calendar/page.jsx", "src/app/Donations/page.jsx", "src/app/About/page.jsx"],
-  "exclude": ["node_modules"]
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "src/app/Resource/[id]/page.jsx",
+    "src/app/Resources/page.jsx",
+    "src/app/layout.tsx",
+    "src/app/page.jsx",
+    "src/app/Members/page.jsx",
+    "src/app/Calendar/page.jsx",
+    "src/app/Donations/page.jsx",
+    "src/app/About/page.jsx"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
